Clear the session cookies by logging out the user in request
No user is actually logging in because this is an API call using OAuth to create new user from Firefox Credential Manager. However the complete_social_login function logs the user in setting cookies as if the user is logged in. To handle this logout is called to prevent cookies from being set.