Refactor plan data
Refactor plan data to reduce duplication and make the public / private
interfaces clearer.

_STRIPE_PLAN_DATA has a copy of the key data from Stripe
_RELAY_PLANS_BY_COUNTRY_AND_LANGUAGE has the default language for each
country, as well as the Stripe data associated with each country /
language combo.

The structures used by the code are assembled from this data, with
settings overrides applied, and cached for quicker usage on the next
call.