Add map of languages to country codes
Use the CLDR42 Supplimental Data, Language-Territory Information, to
determine the top country by population that speaks each language, for
the best guess of the country.

Use the Firefox default Accept-Language headers as test cases, and to
limit the list of language-country mappings.

This fixes some cases where the language code does not equal the
lowercased country code, such as Danish (da) to Denmark (dk).

Some additional overrides are provided for special cases in the Firefox
data. such as ca-Valencia (Catalan with Valencia script, probably means
Valencian in Span) and gl-gl (Galician in Greenland, probably means
Galaciain in the Galicia region of Spain).