Use uppercase ISO 3166 codes for countries
Change our lowercased country codes to uppercase ISO 3166 Alpha-2 codes.
This helps distinguish them from lowercase ISO 639-1 language codes. For
example, "DE" is the code for the country of Germany, and "de" is the
code for German.

The front-end code already uppercases country codes when comparing, so
the only change is to sync the mock data with the expected data.