Create forwarded email by replacing content
Load the original email as a Python EmailMessage, and then use the
library functions to delete unused headers, replace other headers, wrap
any HTML content, and prefix any text content.

This preserves the MIME boundaries and structure of the existing
message. It also preserves the MIME sections for inline images.