Add expected emails for free users replying
When a free user replies to an email, we send them an email saying we
forwarded the first email, or did not forward future emails.

The tests used just the local portion of a Relay address, which was
clearer in the email output. It now uses the full address with domain.