Clarify regions and SubPlat
The Subscription Platform does not detect the user's region. This means
Relay is responsible for detecting the region, mapping it to the
relevant plan, and forwarding the user to the correct SubPlat page.

SubPlat does have a list of allowed regions, shared with Stripe. Stripe
should check the region assoicated with a user's payment, and reject it
if it is not allowed.