Change CSP style-src hashes for next.js 14
The CSS style files generated by next.js have a sourcemap comment that
is changed by the browser when it is inserted as an inline style. This
means the hash digest no longer matches the file.

The file source ends in  "css.map*/"
The inline style ends in "css.map */", adding a space before the
comment.

This change from tag 2023.11.08.8 adds a second hash for the inline
version of the content. It makes further changes:

* Refactor the code to use pathlib.Path
* Switch to sha256 hash digests, to match the CSP error messages from
  Chrome and to limit the header size increase from doubling the hash
  digests.
* Add a hash digest for the empty string. This avoids a spurious CSP
  error when next.js injects an empty <style> element, before adding the
  content.