Fix CSS sourceMappingURL, add cache invalidation
Django 4.1 started "fixing" the CSS sourceMappingURL comment, which
added an additional space to next.jss CSS files. This meant the content
changed without the name changing, causing issues when caches returned
old version of the file.

This leans into that change, further fixing the URL so that it is
accurate when next.js injects the content as an inline style.

It also adds a short comment to globals.scss and tokens.scss that Sass
will preserve in the output CSS. This allows changing the content of the
CSS, so next.js will change the CSS filenames, and the cache will be
invaldiated. This should only be necessary around Django updates
for django.contrib.staticfiles.