Add AccountMiddleware if available
The CircleCI migrations tests are supposed to emulate production during
a deploy (old code running against the new migrations), but it also
includes ia requirements to the new code (and the related migrations).
This settings change adds the middleware if it is available, and if it
is available, it is also required.