Update references to moved Protocol SCSS file
It looks like it moved, and with their switch to @use over @import,
duplicate imports would cause errors - so we had to switch to @use
for those as well.

We'll eventually have to switch to @import everywhere, since Sass
is looking to deprecate the latter, but for now, this should at
least bring us up-to-date with Protocol again.