Load GA only after checking for enabled metrics
Use a singleton hook useGoogleAnalytics, default false, to control all
the code that emits Google Analytics (GA) events. MyApp watches for
metrics to load, and then, if enabled, runs initGoogleAnalytics to load
the GA script, then set useGoogleAnalytics to true to refresh with the
enabled event code.

This updates useIsLoggedIn to return a string, 'unknown' while loading,
and 'logged-in' / 'logged-out' once the profile API call is processed to
determine the logged-in status. In a similar way, useMetrics returns
'unknown', 'enabled', and 'disabled'. This protects hooks from being
called unnecessarily while the profile API is called and processed.