Add draft metrics, glean_parser, and gen'ed code
The initial metrics for measuring mask usage are in
telemetry/glean/relay-server-metrics.yaml. This new folder can be used
for front-end Glean metrics and other metrics specifications.

glean_parser is used to generate privaterelay/glean/server_events.py and
the empty __init__.py.  The generated file is post-processed with black.
A function in .circleci/python_job.bash has the generation commands and
the post-processing commands. This is used by a new job to validate in
CircleCI that the generated file was not hand-modified. It can also be
used to regenerate the file locally (tested on Linux and macOS).