Handle order issue with useGoogleAnalytics
In some cases (such as npn dev:mocked, or occasionally in the browser),
the function initGoogleAnalytics is run before useGoogleAnalytics has
replaced globalEnableGoogleAnalytics. This changes
globalEnableGoogleAnalytics to start with a default of null. When it is
not yet a function, gaIsInitialized is changed instead, to provide a
default of True for the singletonHook.