Process messages in a multiprocessing pool
multiprocessing runs the target function in a subprocess rather than a
thread. This is slower to start up than a thread, and requires
django.setup() to initialize the application. However, it does allow us
to terminate a stuck process, which happens sometimes in email
processing and frequently in DLQ processing.