Use PR user login to detect dependabot[bot]
According to a recent article by synacktiv, the Dependabot credentials
can be hijacked and used to create new jobs with:
`github.actor == 'dependabot[bot]'`
A way to check if dependabot is the real origin is to instead check:
`github.event.pull_request.user.login == 'dependabot[bot]'`

This check is repeated by the "dependabot/fetch-metadata@v2" code, so
Relay has a partial mitigation. This change will more completely
mitigate the issue.