Switch metrics_event to metrics and logs
The send_ga_ping code is sending data to the deprecated Universal
Analytics endpoint, which drops the data. Instead, send the core data as
a tagged counter metric, and as a log message. The log includes a
truncated hash of the GA UUID, to allow estimation of unique users. The
metric does not include the GA UUID, to avoid high-cardinality issues.