Update complaint message to match live complaint
The "source" was the mask, but AWS SES uses the generic Reply-To address.
The messageId is set, probably by SES when sending. The null values in
commonHeaders are not present.

These changes break the mask disable test, which uses the "source". It
is marked as an expected failure for now.