MPP-3927: Pre-fetch profile for domain address
The API for domain address uses the subdomain on the Profile to
construct the full domain address. Previously, this required 1 query to
get all the domain addresses, then 2 queries _per address_ to get the
subdomain. So, a user with 100 domain addresses would use 201 queries to
get their list of addresses when they visited the dashboard.

With this change, the user and profile query is done once, so a user
needs 1 query if they have no domain addresses, and 3 total queries for
any number of domain addresses.

The relay addresses API does not use user or profile data, so it
continues to use 1 query for any number of addresses.