Fix metrics prefix
The metrics configuration used `statsd_prefix` instead of the correct
`statsd_namespace` for the datadog backend. Instead, a prefix was added
to all metrics with `metrics = markus.get_metrics("fx-private-relay")`,
with dashes turned into dots.

This corrects the configuration, removes the static prefix, and fixes
the tests that used the previous static prefix.